+++ src/workers/gcodeEngine.ts
@@ -22,6 +22,10 @@ import {
   nearestNeighbourSort,
   joinSubpaths,
 } from "./gcodeEngine/stages/pathOptimization";
+import {
+  clipSubpathsToRect,
+  clipSubpathsToBed,
+} from "./gcodeEngine/stages/clipping";
 
 // ── Types ─────────────────────────────────────────────────────────────────────
 
@@ -211,124 +215,7 @@ export function flattenToSubpaths(
 
 // ── Bed clipping ─────────────────────────────────────────────────────────────
 
-/**
- * Liang-Barsky line-segment clipper.
- * Clips segment (x0,y0)→(x1,y1) against axis-aligned rectangle.
- * Returns [t0, t1] parametric range of the visible portion, or null if
- * the segment lies entirely outside the rectangle.
- */
-function liangBarsky(
-  x0: number,
-  y0: number,
-  x1: number,
-  y1: number,
-  xMin: number,
-  xMax: number,
-  yMin: number,
-  yMax: number,
-): [number, number] | null {
-  const dx = x1 - x0;
-  const dy = y1 - y0;
-  let t0 = 0,
-    t1 = 1;
-  const tests: [number, number][] = [
-    [-dx, x0 - xMin],
-    [dx, xMax - x0],
-    [-dy, y0 - yMin],
-    [dy, yMax - y0],
-  ];
-  for (const [p, q] of tests) {
-    if (p === 0) {
-      if (q < 0) return null; // parallel and outside
-    } else {
-      const r = q / p;
-      if (p < 0) t0 = Math.max(t0, r);
-      else t1 = Math.min(t1, r);
-      if (t0 > t1) return null;
-    }
-  }
-  return [t0, t1];
-}
-
-/**
- * Clips an array of polyline subpaths against an explicit axis-aligned
- * rectangle [xMin, xMax] × [yMin, yMax] (all in machine mm).
- * A single input subpath may split into several output subpaths when its
- * segments exit and re-enter the rectangle.  Segments entirely outside are
- * silently dropped.
- */
-export function clipSubpathsToRect(
-  subpaths: Subpath[],
-  xMin: number,
-  xMax: number,
-  yMin: number,
-  yMax: number,
-): Subpath[] {
-  const result: Subpath[] = [];
-
-  for (const subpath of subpaths) {
-    let current: Subpath = [];
-
-    const interp = (p0: Pt, p1: Pt, t: number): Pt => ({
-      x: p0.x + t * (p1.x - p0.x),
-      y: p0.y + t * (p1.y - p0.y),
-    });
-
-    for (let i = 0; i + 1 < subpath.length; i++) {
-      const p0 = subpath[i];
-      const p1 = subpath[i + 1];
-      const clip = liangBarsky(p0.x, p0.y, p1.x, p1.y, xMin, xMax, yMin, yMax);
-
-      if (clip === null) {
-        // Segment fully outside — flush any open run.
-        if (current.length >= 2) result.push(current);
-        current = [];
-        continue;
-      }
-
-      const [t0, t1] = clip;
-
-      if (t0 > 1e-9) {
-        // Segment enters the rect mid-way — flush previous run and start fresh.
-        if (current.length >= 2) result.push(current);
-        current = [interp(p0, p1, t0)];
-      } else if (current.length === 0) {
-        // Starting a new run from inside (or right on) the rect boundary.
-        current.push(p0);
-      }
-      // Append the exit point.
-      current.push(interp(p0, p1, t1));
-
-      if (t1 < 1 - 1e-9) {
-        // Segment exits the rect before reaching p1 — flush this run.
-        if (current.length >= 2) result.push(current);
-        current = [];
-      }
-    }
-
-    if (current.length >= 2) result.push(current);
-  }
-
-  return result;
-}
-
-/**
- * Clips an array of polyline subpaths (already in machine mm coordinates)
- * against the bed rectangle implied by config.  A single input subpath may
- * split into several output subpaths when its segments exit and re-enter the
- * bed area.  Segments entirely outside the bed are silently dropped.
- */
-export function clipSubpathsToBed(
-  subpaths: Subpath[],
-  config: MachineConfig,
-): Subpath[] {
-  const isCenter = config.origin === "center";
-  const xMin = isCenter ? -config.bedWidth / 2 : 0;
-  const xMax = isCenter ? config.bedWidth / 2 : config.bedWidth;
-  const yMin = isCenter ? -config.bedHeight / 2 : 0;
-  const yMax = isCenter ? config.bedHeight / 2 : config.bedHeight;
-  return clipSubpathsToRect(subpaths, xMin, xMax, yMin, yMax);
-}
+export { clipSubpathsToRect, clipSubpathsToBed };
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
