+++ src/workers/gcodeEngine/stages/pathOptimization.ts
