+++ tests/unit/gcodePathOptimization.test.ts
